Refactored drain into delta time approach, moved timer cancel into health regenerate function, added Distance specific logging category to output that allows verbose logging

+++ Source/Distance/Public/AIEnemy.h
@@ -66,9 +66,8 @@ public:
 	UPROPERTY(EditAnywhere)
 	bool drainLight;
 
-	int drainCounter;
-
-	int drainRate;
+	float drainCounter;
+	float drainRate;
 
 	float speedCounter;
 	float speedLimit;
